Rename image authoring files
