+++ frontend/src/features/authoring/images.ts
